refactor(dvdrental-api): add helpers for nullable pgtype conversions

+++ dev-realm/dvdrental-api/internal/addresses/types.go
@@ -20,7 +20,7 @@ type AddressRequest struct {
 	Address    string  `json:"address"`
 	Address2   *string `json:"address2"`
 	District   string  `json:"district"`
-	CityID     int32   `json:"city_id"`
+	CityID     int16   `json:"city_id"`
 	PostalCode *string `json:"postal_code"`
 	Phone      string  `json:"phone"`
 }
@@ -29,7 +29,7 @@ type UpdateAddressPartialRequest struct {
 	Address    *string `json:"address,omitempty"`
 	Address2   *string `json:"address2,omitempty"`
 	District   *string `json:"district,omitempty"`
-	CityID     *int32  `json:"city_id,omitempty"`
+	CityID     *int16  `json:"city_id,omitempty"`
 	PostalCode *string `json:"postal_code,omitempty"`
 	Phone      *string `json:"phone,omitempty"`
 }
